requirements.txt
flask
python-dotenv
psycopg2-binary
flask-jwt-extended
flask_sqlalchemy
SQLAlchemy-serializer